Move `createSelectorCreator` and `createStructuredSelector` into their own files
This was done because the new overload signature for `createSelectorCreator` was added and the index.ts file was getting a little too crowded.
