+++ src/createStructuredSelector.ts
@@ -1,0 +1,59 @@
+import { createSelector } from './createSelectorCreator'
+
+import type { CreateSelectorFunction } from './createSelectorCreator'
+import type { AnyFunction, Head, ObjValueTuple, Selector, Tail } from './types'
+import type { MergeParameters } from './versionedTypes'
+
+interface SelectorsObject {
+  [key: string]: AnyFunction
+}
+
+export interface StructuredSelectorCreator {
+  <
+    SelectorMap extends SelectorsObject,
+    SelectorParams = MergeParameters<ObjValueTuple<SelectorMap>>
+  >(
+    selectorMap: SelectorMap,
+    selectorCreator?: CreateSelectorFunction<any, any>
+  ): (
+    // Accept an arbitrary number of parameters for all selectors
+    // The annoying head/tail bit here is because TS isn't convinced that
+    // the `SelectorParams` type is really an array, so we launder things.
+    // Plus it matches common usage anyway.
+    state: Head<SelectorParams>,
+    ...params: Tail<SelectorParams>
+  ) => {
+    [Key in keyof SelectorMap]: ReturnType<SelectorMap[Key]>
+  }
+
+  <State, Result = State>(
+    selectors: {
+      [K in keyof Result]: Selector<State, Result[K], never>
+    },
+    selectorCreator?: CreateSelectorFunction<any, any>
+  ): Selector<State, Result, never>
+}
+
+// Manual definition of state and output arguments
+export const createStructuredSelector = ((
+  selectors: SelectorsObject,
+  selectorCreator = createSelector
+) => {
+  if (typeof selectors !== 'object') {
+    throw new TypeError(
+      'createStructuredSelector expects first argument to be an object ' +
+        `where each property is a selector, instead received a ${typeof selectors}`
+    )
+  }
+  const objectKeys = Object.keys(selectors)
+  const resultSelector = selectorCreator(
+    objectKeys.map(key => selectors[key]),
+    (...values: any[]) => {
+      return values.reduce((composition, value, index) => {
+        composition[objectKeys[index]] = value
+        return composition
+      }, {})
+    }
+  )
+  return resultSelector
+}) as unknown as StructuredSelectorCreator
